Create new test file, update index supported rst from Y to P, test not implemented bins
